Assume HWND is non-null
The `Option<HWND>` stems from an earlier time when a Window could shut
itself down from the Win32 side. We now force all window shutdowns to
happen from the top down by dropping the `Window` object so we can
safely assume that the HWND is non-null for the lifetime of the
`Window`.